+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Changed
 ### Fixed
+- Fixed potential buffer overflow error in `bus_send_message()`
 
 ## [0.0.4] - 2026-01-31