+++ changelog.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Always set the `charset=UTF-8` on web responses.
 - Fixed upload of the firmware and partitioning
 - Fixed mark the new volume as valid on successful boot
+- Auto reload the `/update` page on successful update.
 
 ## [0.0.4] - 2026-01-31